rename app to index and index to entry
