+++ packages/desktop/index.html
@@ -19,6 +19,6 @@
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root" class="flex flex-col h-dvh"></div>
     <div data-tauri-decorum-tb class="w-0 h-0 hidden" />
-    <script src="/src/index.tsx" type="module"></script>
+    <script src="/src/entry.tsx" type="module"></script>
   </body>
 </html>
