+++ packages/desktop/src/entry.tsx
@@ -1,0 +1,5 @@
+if (location.pathname === "/loading") {
+  import("./loading")
+} else {
+  import("./")
+}
